backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Sistema de Gestión Académica"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "uvicorn>=0.44.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
]
